+++ tests/Traits/PdfAttachmentTraitTest.php
@@ -14,13 +14,14 @@ declare(strict_types=1);
 namespace fpdf\Tests\Traits;
 
 use fpdf\Enums\PdfDestination;
+use fpdf\Enums\PdfVersion;
 use fpdf\PdfException;
 use fpdf\Tests\fixture\PdfDocumentAttachment;
 use PHPUnit\Framework\TestCase;
 
 class PdfAttachmentTraitTest extends TestCase
 {
-    public function testAttachment(): void
+    public function testAttachmentWithDescription(): void
     {
         $file = __DIR__ . '/../resources/attachment.txt';
 
@@ -29,6 +30,19 @@ class PdfAttachmentTraitTest extends TestCase
         $document->addAttachment(file: $file, desc: 'Attached File.');
         $document->output(PdfDestination::STRING);
         self::assertSame(1, $document->getPage());
+        self::assertSame(PdfVersion::VERSION_1_6, $document->getPdfVersion());
+    }
+
+    public function testAttachmentWithoutDescription(): void
+    {
+        $file = __DIR__ . '/../resources/attachment.txt';
+
+        $document = new PdfDocumentAttachment();
+        $document->addPage();
+        $document->addAttachment(file: $file);
+        $document->output(PdfDestination::STRING);
+        self::assertSame(1, $document->getPage());
+        self::assertSame(PdfVersion::VERSION_1_4, $document->getPdfVersion());
     }
 
     public function testInvalidFile(): void
@@ -48,5 +62,6 @@ class PdfAttachmentTraitTest extends TestCase
         $document->addPage();
         $document->output(PdfDestination::STRING);
         self::assertSame(1, $document->getPage());
+        self::assertSame(PdfVersion::VERSION_1_3, $document->getPdfVersion());
     }
 }